upload

+++ aquaLib/easyCommands/commands/clear.js
@@ -10,18 +10,19 @@ module.exports = (Client, message) => {
     numberToDelete = Number(args[1])
     if (config.autoCallMessageDelete == true && message.deletable == true) message.delete()
     if (args[1]) {
-        if(args[1] == "all") {
-            return message.channel.messages.fetch({ limit: 100 }).then(messages => {
-                if(messages.size == 0) return sendMessage('error', "Il n'y a aucun messages dans ce salon.", message.channel)
-                message.channel.send({ embeds: [new Discord.EmbedBuilder().setDescription("Tant que le bot ne s'arrète pas, la suppression continue automatiquement.").setTimestamp().setColor(app.botColor).setTitle("L'action à bien été lancée, la suppréssion peut durer plusieurs minutes").setFooter({ text: "Par " + app.name, iconURL: app.iconURL })]})
-                setTimeout(() => {
-                    messages.forEach(msg => {
-                        msg.delete().then(() => {}).catch(err => console.error(err.toString() + "\nCette erreur n'est pas importante"))
-                    })
-                }, 1000)
-            })
-        }
         if (!isNaN(args[1]) && args[1] >= 1 && args[1] <= 100) {
+            if (args[2] == "true") {
+                return message.channel.messages.fetch({ limit: args[1] }).then(messages => {
+                    if (messages.size == 0) return sendMessage('error', "Il n'y a aucun messages dans ce salon.", message.channel)
+                    message.channel.send({ embeds: [new Discord.EmbedBuilder().setDescription("Tant que le bot ne s'arrète pas, la suppression continue.\nDans une suppression forcée, les messages sont supprimés un par un.").setTimestamp().setColor(app.botColor).setTitle("L'action à bien été lancée, la suppréssion peut prendre du temps").setFooter({ text: "Par " + app.name, iconURL: app.iconURL })] }).then(msg => { if (config.BotMessageDelete.autoBotMessageDelete == true) { setTimeout(() => { if (msg.deletable) msg.delete().catch(err => console.error(err + "\nCette erreur n'est pas importante")) }, config.BotMessageDelete.timeout) } })
+                    setTimeout(() => {
+                        messages.forEach(msg => {
+                            msg.delete().then(() => { }).catch(err => console.error(err.toString() + "\nCette erreur n'est pas importante"))
+                        })
+                    }, 5000)
+                })
+            }
+            if(args[2] && args[2] !== "false") return sendMessage('error', "Le paramètre Force n'accepte qu'un boolean (true ou false)", message.channel)
             for (let i = args[1].split("").length; i != 0; i--) { if (args[1].split("")[i] == ".") return sendMessage('error', "Veuillez indiquer un nombre entier.", message.channel) }
             message.channel.bulkDelete(numberToDelete, true).then(() => {
                 sendMessage("success", `Vous avez supprimé ${args[1]} message(s).`, message.channel)
